+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Extended predict.w2v with nearest if you pass on a vector or matrix. This allows to perform word2vec analogies or extract other similarities.
 - Added word2vec_similarity
+- Change classes returned by word2vec to 'word2vec_trained' and read.word2vec to 'word2vec'
+- Add detailed docs of predict.word2vec and as.matrix.word2vec
 
 ## CHANGES IN word2vec VERSION 0.1.0